+++ adbyss_psl/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 
 
+## [0.23.0](https://github.com/Blobfolio/adbyss/releases/tag/v0.23.0) - 2025-11-16
+
+### New
+
+* `Domain::ends_with`
+* `Domain::eq` (const alternative)
+* `Domain::eq_tld`
+
+### Changed
+
+* Bump MSRV `1.91`
+* Make `sqlx` traits generic over `sqlx::Database`
+* Update suffix database
+
+### Breaking
+
+* Rename `sqlx-mysql` feature `sqlx`
+
+
+
 ## [0.22.0](https://github.com/Blobfolio/adbyss/releases/tag/v0.22.0) - 2025-10-30
 
 ### Changed